Lab3_CodeRevGithub/Lab3S01/requirements.txt
pandas~=2.2.3
numpy~=2.2.4
scipy~=1.15.2
tqdm~=4.67.1
PyGithub~=2.6.1
requests~=2.32.3